// src/interfaces/genres.ts
export enum Genre {
  Chillhop = "chillhop",
  ContemporaryCountry = "contemporary country",
  Broadway = "broadway",
  DancePop = "dance pop",
  AnthemWorship = "anthem worship",
  EmoRap = "emo rap",
  AlbumRock = "album rock",
  Hollywood = "hollywood",
  BigRoom = "big room",
  AlternativeRB = "alternative r&b",
}